Fix the way save method is called
I initially called storage.save which would only call save on the storage
engine used. It introduced a bug of not saving instance of a class into the
storage engine. so i fixed it by calling BaseModel.save which does the job of adding
the instance and also saving to the storage engine used